src/requirements.txt
Django==4.1
django-tinymce==3.4.0
Pillow==9.2.0
